@@ -1,7 +1,5 @@
-'use strict';
+import mm from "minimatch";
 
-const mm = require('minimatch');
-
-console.time('minimatch');
-console.log(mm.makeRe('**/*').test('foo/bar/baz/qux.js'));
-console.timeEnd('minimatch');
+console.time("minimatch");
+console.log(mm.makeRe("**/*").test("foo/bar/baz/qux.js"));
+console.timeEnd("minimatch");
