@@ -1,9 +1,7 @@
-'use strict';
+import pm from "../index.js";
 
-const pm = require('..');
-
-console.log(pm.makeRe('*'));
+console.log(pm.makeRe("*"));
 // /^(?:(?!\.)(?=.)[^\\\/]*?\/?)$/
 
-console.log(pm.makeRe('*', { dot: true }));
+console.log(pm.makeRe("*", { dot: true }));
 // /^(?:(?!\.{1,2}(?:\/|$))(?=.)[^\\\/]*?\/?)$/
