@@ -1,8 +1,6 @@
-'use strict';
+import pm from "../index.js";
 
-const pm = require('..');
-
-const isMatch = pm.matcher('*', { ignore: 'f*' });
-console.log(isMatch('foo')); //=> false
-console.log(isMatch('bar')); //=> true
-console.log(isMatch('baz')); //=> true
+const isMatch = pm.matcher("*", { ignore: "f*" });
+console.log(isMatch("foo")); //=> false
+console.log(isMatch("bar")); //=> true
+console.log(isMatch("baz")); //=> true
