@@ -1,57 +1,56 @@
-/*global navigator*/
-'use strict';
-
-const {
+import {
   REGEX_BACKSLASH,
   REGEX_REMOVE_BACKSLASH,
   REGEX_SPECIAL_CHARS,
-  REGEX_SPECIAL_CHARS_GLOBAL
-} = require('./constants');
+  REGEX_SPECIAL_CHARS_GLOBAL,
+} from "./constants.js";
 
-exports.isObject = val => val !== null && typeof val === 'object' && !Array.isArray(val);
-exports.hasRegexChars = str => REGEX_SPECIAL_CHARS.test(str);
-exports.isRegexChar = str => str.length === 1 && exports.hasRegexChars(str);
-exports.escapeRegex = str => str.replace(REGEX_SPECIAL_CHARS_GLOBAL, '\\$1');
-exports.toPosixSlashes = str => str.replace(REGEX_BACKSLASH, '/');
+export const isObject = (val) =>
+  val !== null && typeof val === "object" && !Array.isArray(val);
+export const hasRegexChars = (str) => REGEX_SPECIAL_CHARS.test(str);
+export const isRegexChar = (str) => str.length === 1 && hasRegexChars(str);
+export const escapeRegex = (str) =>
+  str.replace(REGEX_SPECIAL_CHARS_GLOBAL, "\\$1");
+export const toPosixSlashes = (str) => str.replace(REGEX_BACKSLASH, "/");
 
-exports.isWindows = () => {
-  if (typeof navigator !== 'undefined' && navigator.platform) {
+export const isWindows = () => {
+  if (typeof navigator !== "undefined" && navigator.platform) {
     const platform = navigator.platform.toLowerCase();
-    return platform === 'win32' || platform === 'windows';
+    return platform === "win32" || platform === "windows";
   }
 
-  if (typeof process !== 'undefined' && process.platform) {
-    return process.platform === 'win32';
+  if (typeof process !== "undefined" && process.platform) {
+    return process.platform === "win32";
   }
 
   return false;
 };
 
-exports.removeBackslashes = str => {
-  return str.replace(REGEX_REMOVE_BACKSLASH, match => {
-    return match === '\\' ? '' : match;
+export const removeBackslashes = (str) => {
+  return str.replace(REGEX_REMOVE_BACKSLASH, (match) => {
+    return match === "\\" ? "" : match;
   });
 };
 
-exports.escapeLast = (input, char, lastIdx) => {
+export const escapeLast = (input, char, lastIdx) => {
   const idx = input.lastIndexOf(char, lastIdx);
   if (idx === -1) return input;
-  if (input[idx - 1] === '\\') return exports.escapeLast(input, char, idx - 1);
+  if (input[idx - 1] === "\\") return escapeLast(input, char, idx - 1);
   return `${input.slice(0, idx)}\\${input.slice(idx)}`;
 };
 
-exports.removePrefix = (input, state = {}) => {
+export const removePrefix = (input, state = {}) => {
   let output = input;
-  if (output.startsWith('./')) {
+  if (output.startsWith("./")) {
     output = output.slice(2);
-    state.prefix = './';
+    state.prefix = "./";
   }
   return output;
 };
 
-exports.wrapOutput = (input, state = {}, options = {}) => {
-  const prepend = options.contains ? '' : '^';
-  const append = options.contains ? '' : '$';
+export const wrapOutput = (input, state = {}, options = {}) => {
+  const prepend = options.contains ? "" : "^";
+  const append = options.contains ? "" : "$";
 
   let output = `${prepend}(?:${input})${append}`;
   if (state.negated === true) {
@@ -60,11 +59,11 @@ exports.wrapOutput = (input, state = {}, options = {}) => {
   return output;
 };
 
-exports.basename = (path, { windows } = {}) => {
-  const segs = path.split(windows ? /[\\/]/ : '/');
+export const basename = (path, { windows } = {}) => {
+  const segs = path.split(windows ? /[\\/]/ : "/");
   const last = segs[segs.length - 1];
 
-  if (last === '') {
+  if (last === "") {
     return segs[segs.length - 2];
   }
 
